requirements.txt
hydra-core==1.3.2
librosa==0.9.2
lightning==2.1.2
matplotlib==3.1.2
musdb==0.4.0
museval==0.4.0
numpy==1.17.4
omegaconf==2.3.0
openunmix==1.2.1
pandas==1.3.4
pytorch_lightning==2.1.2
PyYAML==6.0.1
scikit_learn==1.3.2
torch==1.12.1
torchaudio==0.12.1
torchmetrics==1.2.1
tqdm==4.62.3
